refactor: 훅으로 분리
